Added comments for ResponseUtil class and removed redundant comments
